+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2023-08-08
+
+### Added
+
+- alldebrid support
+- independent simultaneous rclone_RD mounts 
+- plex_debrid settings.json to automatically include alldebrid and realdebrid as debrid services
+
+### Changed
+
+- healthcheck process to monitor both alldebrid and realdebrid mounts 
+- duplicate cleanup process to use alldebrid and realdebrid mounts
+
+### Removed
+
+- duplicate format_time function
+
+
 ## [1.3.1] - 2023-07-27
 
 ### Changed